+++ crates/rsvelte_core/src/svelte2tsx/template/attributes/mod.rs
@@ -1,3 +1,5 @@
 //! Attribute and directive emission.
 
+pub(super) mod action;
+pub(super) mod let_;
 pub(super) mod svg;
